Fix : changed average_nb_of_day column type from int to float

+++ migrations/20220830090457_add_average_nb_of_days_to_users.ts
@@ -1,7 +1,7 @@
 
 export async function up(knex) {
     return knex.schema.table('users', (table) => {
-        table.integer('average_nb_of_days')
+        table.float('average_nb_of_days')
     })
 }
